Add login functionality and error handling. Fixed entitynotfound error handling.

+++ demo/src/main/java/com/example/demo/controller/MoonController.java
@@ -1,5 +1,6 @@
 package com.example.demo.controller;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.slf4j.Logger;
@@ -18,6 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.example.demo.entities.Moon;
 import com.example.demo.exceptions.EntityNotFound;
+import com.example.demo.exceptions.NotAuthorizedException;
 import com.example.demo.service.MoonService;
 import org.springframework.web.bind.annotation.RequestParam;
 
@@ -29,41 +31,68 @@ public class MoonController {
     @Autowired
     private MoonService moonService;
 
+    @Autowired
+    private AuthenticateController authenticateController;
+
     @ExceptionHandler(EntityNotFound.class)
     public ResponseEntity<String> entityNotFound(EntityNotFound e) {
         moonLogger.error(e.getLocalizedMessage(), e);
         return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
     }
 
+    @ExceptionHandler
+    public ResponseEntity<String> notAuthorized(NotAuthorizedException e) {
+        moonLogger.error(e.getLocalizedMessage(), e);
+        return new ResponseEntity<>(e.getMessage(), HttpStatus.UNAUTHORIZED);
+    }
+
     @GetMapping("/api/moons")
     public ResponseEntity<List<Moon>> findAll() {
+        if (!authenticateController.isLoggedIn()) {
+            throw new NotAuthorizedException("You must login to perform this action.");
+        }
         return new ResponseEntity<>(moonService.findAll(), HttpStatus.OK);
     }
 
     @GetMapping("/api/moon/id/{id}")
     public ResponseEntity<Moon> findById(@PathVariable int id) {
+        if (!authenticateController.isLoggedIn()) {
+            throw new NotAuthorizedException("You must login to perform this action.");
+        }
         return new ResponseEntity<>(this.moonService.findById(id), HttpStatus.OK);
     }
 
     @GetMapping("/api/moon/{name}")
     public ResponseEntity<Moon> findByName(@PathVariable String name) {
+        if (!authenticateController.isLoggedIn()) {
+            throw new NotAuthorizedException("You must login to perform this action.");
+        }
         return new ResponseEntity<>(this.moonService.findByName(name), HttpStatus.OK);
     }
 
     @PostMapping("/api/moon")
     public ResponseEntity<String> createMoon(@RequestBody Moon newMoon) {
+        if (!authenticateController.isLoggedIn()) {
+            throw new NotAuthorizedException("You must login to perform this action.");
+        }
         String message = this.moonService.createMoon(newMoon);
         return new ResponseEntity<>(message, HttpStatus.CREATED);
     }
 
     @DeleteMapping("/api/moon/id/{id}")
     public ResponseEntity<String> deleteMoon(@PathVariable int id) {
+        if (!authenticateController.isLoggedIn()) {
+            throw new NotAuthorizedException("You must login to perform this action.");
+        }
         String message = this.moonService.deleteById(id);
         return new ResponseEntity<>(message, HttpStatus.OK);
     }
 
     @GetMapping("/api/planet/moons/{ownerId}")
     public ResponseEntity<List<Moon>> findByOwnerId(@PathVariable int ownerId) {
+        if (!authenticateController.isLoggedIn()) {
+            throw new NotAuthorizedException("You must login to perform this action.");
+        }
         return new ResponseEntity<>(moonService.findByOwnerId(ownerId), HttpStatus.OK);
     }
